requirements.txt
pytorch
pytorch-transformers
pytorch_pretrained_bert
flask-cors
tqdm
